Optimization on experience increase/decrease
- Change level to Integer (less cast)
- Add JSON support
- Little adjustment on Cursed Weapons fix (instance problem)
- Add config for max player/pet level
- Last write on StopRotation need to be int not byte

+++ src/main/java/com/l2jserver/gameserver/model/actor/stat/NpcStat.java
@@ -28,7 +28,7 @@ public class NpcStat extends CharStat
 	}
 	
 	@Override
-	public byte getLevel()
+	public int getLevel()
 	{
 		return getActiveChar().getTemplate().getLevel();
 	}
